# docs/swagger.yaml
definitions:
  main.Task:
    properties:
      description:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /tasks:
    get:
      consumes:
      - application/json
      description: Возвращает массив задач
      produces:
      - application/json
      responses:
        "200":
          description: Массив задач
          schema:
            items:
              $ref: '#/definitions/main.Task'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создает задачу, данные которой передаются в теле запроса,
      parameters:
      - description: Данные новой задачи
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/main.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Созданная задача
          schema:
            $ref: '#/definitions/main.Task'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создает задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет задачу по ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удаление задачи
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Обновляет данные задачи по ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Данные задачи для обновления
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/main.Task'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная задача
          schema:
            $ref: '#/definitions/main.Task'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление задачи
      tags:
      - tasks
swagger: "2.0"